Split up emacs initilization into seprate files.
Seporated the configurations for emacs so that the files were cleaner
and easier to find things in. Still have possible reorganizing to do
such as possible moving the package initialization config values to
proper spots, but I haven't decided which is cleaner yet. For now this
is much better than before.