Reduced qutebrowser config
The qutebrowser config had all the settings possible for qute browswer
to use present in the config. However, as qutebrowser evolves (or as
different versions get used across different devices) I fond it annoying
to have to constantly be updating settings I never changed from the
defaults. Now, the config only contains the settings that are different
from default. This also allows any quality of life updates to the
defaults to take affect.