Zsh - Source setup files prior to loading omz
There are some enviornment variables that get used or created during the
loading of Oh-My-Zshell which should be set beforehand. Because I want
to have these conveniently set in my shell setup folder, I've decided to
just source the folder before and after so that all values are the way I
want them to be post omz setup.