Emacs - Rework org work directory usage
In the future, I may want to separate my personal org files from my work
org files so here I'm creating a direct reference to my org work
directory that gets referenced anywhere that I want to access work files.

This also substitutes the usage of `file-name-as-directory` with the
cleaner `file-name-concat` since that allows me to just list a system
file path and have it be built as the system expects.