Emacs - Fix folding issue
The fold module in doom emacs was causing some troubles for me so I
rewrote the folding logic that I wanted in place. Now, the folding
commands will properly trigger from evil mode's fold function hierarchy.

This change also sets the tab button to properly behave as a fold button
when not in insert mode.