Emacs - Updated tab behavior
Pulled the tab behavior that is defined within the doom core and then
tweaked it to add my folding desires. With how I had tab set up earlier,
it would only fold and wouldn't trigger snippets the way it needed to.
Although I don't like pulling out this hard-coded value, it works cleanly
and I don't see an alternative for the time being.