Emacs - Unpin LSP mode for proper breadcrumbs
On the pinned LSP mode commit, there is a bug that causes the lsp mode
breadcrumb headerline to display the same headerline across all windows
which was super annoying. Unpinning it fixed that issue and allows LSP
mode to remain up to date with whatever else may have been fixed.