Emacs - fix config
In some recent updates, the behavior of some packages have changed which
is causing trouble.

- For some reason, global git gutter mode is no longer a thing. I'm not
sure why or even where it was from originally, but I need to look into
it later.

- global tree sitter mode now causes all of emacs to fail when there's
no grammar installed which is terrible behavior. I need to look to see
if there's some way around it, but this took ages to find.