Shell - Don't alias open when it already exists
Open is a command that opens a given file in an external program. On
linux, I need to alias this command to use a 3rd party executable since
open isn't a default command. However, on Macs, open is a command that
already exists and works well. The bit of logic here will just prevent
open from being re-defined if it already exists in the environment.