Bin - Add script to analyze file coupling
When files are edited together regularly within code commits, it likely
means they have a coupling to them. This isn't necessarily bad since
they may be test files or the like, but sometimes it means "separate"
pieces aren't really all that separate. This script will identify all
file pairs that are modified in a commit history and some stats to help
a user identify possible coupling.